Fixes API Gateway probes and deployment script
Updates the liveness and readiness probes for the API Gateway to use the `/api` path.

Removes the `fsGroup` from the security context in `values.yaml` and `values-production.yaml` as it belongs in `podSecurityContext`.

Replaces the `deploy-minikube-fixed.ps1` script, and renames it to `deploy-minikube.ps1`.
The updated script enhances the deployment process, incorporating features such as:
- Clean install option
- Ingress skipping
- Timeout configuration
- Improved error handling
- Docker environment configuration
- Image building within Minikube's environment.
- Retry mechanism for Helm deployments to address webhook errors
- Fallback deployment strategy without Ingress in case of Ingress-related issues

Adds a `fix-ingress-webhook.ps1` script to resolve Ingress Controller webhook issues
by performing a clean re-installation with minimal configuration.